Correct 100 for INT_MIN

+++ pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@ int _atoi(char *s)
 {
 	char *p = s;
 	int sign = 1;
-	int result = 0;
+	long result = 0;
 
 	while (*p < '0' || *p > '9')
 	{
@@ -30,5 +30,5 @@ int _atoi(char *s)
 		result = result * 10 + (*p - '0');
 		p++;
 	}
-	return (sign * result);
+	return (int)(sign * result);
 }
